# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Awards Management API",
        "description": "API for managing awards, nominees, voting, and events",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Awards Management API Server"
        }
    ],
    "paths": {
...
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get list of categories",
                "description": "Returns list of categories with filtering options",
                "operationId": "getCategories",
                "parameters": [
                    {
                        "name": "region",
                        "in": "query",
                        "description": "Filter by region",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Global",
                                "Americas",
                                "Europe",
                                "Africa",
                                "Asia-Pacific",
                                "Middle East"
                            ]
                        }
                    },
                    {
                        "name": "voting_only",
                        "in": "query",
                        "description": "Show only categories with voting open",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Best Innovation Award"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Award for outstanding innovation"
                                                    },
                                                    "criteria": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "region": {
                                                        "type": "string",
                                                        "example": "Global"
                                                    },
                                                    "current_nominees": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "voting_open": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "voting_ends_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "days_remaining": {
                                                        "type": "integer",
                                                        "example": 15
                                                    },
                                                    "is_voting_active": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "color": {
                                                        "type": "string",
                                                        "example": "#f0f9ff"
                                                    },
                                                    "icon": {
                                                        "type": "string",
                                                        "example": "trophy"
                                                    },
                                                    "image_url": {
                                                        "type": "string",
                                                        "example": "https://example.com/image.jpg"
                                                    },
                                                    "total_votes": {
                                                        "type": "integer",
                                                        "example": 150
                                                    },
                                                    "nominees_count": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "winners_count": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get category information",
                "description": "Returns category data with nominees",
                "operationId": "getCategoryById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Best Innovation Award"
                                                },
                                                "description": {
                                                    "type": "string"
                                                },
                                                "criteria": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "region": {
                                                    "type": "string"
                                                },
                                                "voting_open": {
                                                    "type": "boolean"
                                                },
                                                "voting_starts_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "voting_ends_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "days_remaining": {
                                                    "type": "integer"
                                                },
                                                "is_voting_active": {
                                                    "type": "boolean"
                                                },
                                                "color": {
                                                    "type": "string"
                                                },
                                                "icon": {
                                                    "type": "string"
                                                },
                                                "image_url": {
                                                    "type": "string"
                                                },
                                                "total_votes": {
                                                    "type": "integer"
                                                },
                                                "nominees": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            },
                                                            "organization": {
                                                                "type": "string"
                                                            },
                                                            "description": {
                                                                "type": "string"
                                                            },
                                                            "image_url": {
                                                                "type": "string"
                                                            },
                                                            "votes": {
                                                                "type": "integer"
                                                            },
                                                            "voting_percentage": {
                                                                "type": "number",
                                                                "format": "float"
                                                            },
                                                            "can_vote": {
                                                                "type": "boolean"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
...
    },
...
}